requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.0.5
djongo==1.3.4
dnspython==2.1.0
idna==2.10
joblib==1.0.1
lxml==4.6.3
ntlm-auth==1.5.0
numpy==1.20.2
oauthlib==3.1.0
pandas==1.2.4
plotly==4.14.3
pycparser==2.20
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.2
SharePlum==0.5.1
six==1.15.0
sqlparse==0.2.4
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
